+++ allalgorithms/searches/__init__.py
@@ -1,1 +1,2 @@
 from .binary_search import *
+from .fibonacci_search import *